+++ sdk/maps/Azure.Maps.Search/CHANGELOG.md
@@ -1,5 +1,21 @@
 # Release History
 
+## 2.0.0-beta.5 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 2.0.0-beta.4 (2024-09-30)
+
+### Bugs Fixed
+
+- Fix the issue where `Iso` is always `null` in the `GetReverseGeocoding` response
+
 ## 2.0.0-beta.3 (2024-09-23)
 
 ### Breaking Changes